Types `long` and `unsigned long` should be avoided when either specific size
or pointer-sized is needed.

And then arguably avoid others such as long long.

The best types are in `stdint.h`.
Second past are the NT/EFI types.
There are few places to use long, mainly with the Msvc intrinsics.

This also fixes the two definitions of `NTSTATUS` to be the same. Later we will single instance them.

----
This pull request is a code cleanup that enhances portability by updating legacy integer type definitions to fixed-width types.

The changes standardize integer types across the codebase to improve consistency and support clang cleanup efforts.
- `MsvmPkg/Include/Private/EfiNt.h`: Replaces legacy type definitions with fixed-width types from <stdint.h>, updates include directives, and redefines NTSTATUS.
- `MsvmPkg/Sec/X64/SecP.h`: Changes function return types from "long long" to "int64_t" for better type consistency.
- `MsvmPkg/Include/Vmbus/NtStatus.h`: Updates NTSTATUS definition to use int32_t.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #60275790

+++ MsvmPkg/Include/Vmbus/NtStatus.h
@@ -4,11 +4,10 @@
   Copyright (c) Microsoft Corporation.
   SPDX-License-Identifier: BSD-2-Clause-Patent
 **/
-
 #pragma once
+#include <stdint.h>
 
-
-typedef long NTSTATUS;
+typedef int32_t NTSTATUS;
 #define NT_SUCCESS(Status) (((NTSTATUS)(Status)) >= 0)
 
 //
